+++ Johnson/ExternalAPIs/ITyAndSashaApi.cs
@@ -1,8 +1,10 @@
-﻿namespace Shockah.Johnson;
+﻿using Nickel;
+
+namespace Shockah.Johnson;
 
 public interface ITyAndSashaApi
 {
-	bool IsWild(Card card, State s, Combat c);
-
 	Deck TyDeck { get; }
+
+	ICardTraitEntry WildTrait { get; }
 }